Consolidate file_utils image tests
