+++ tests/co_op_translator/utils/common/test_file_utils.py
@@ -1,4 +1,5 @@
 from pathlib import Path
+import tempfile
 import pytest
 from co_op_translator.utils.common.file_utils import (
     read_input_file,
@@ -194,18 +195,6 @@ def test_delete_translated_markdown_files(temp_dir):
     assert not ko_dir.exists()
 
 
-def test_delete_translated_images(temp_dir):
-    """Test deleting translated images."""
-    image_dir = temp_dir / "images"
-    image_dir.mkdir()
-    (image_dir / "test.ko.png").touch()
-    (image_dir / "test.en.png").touch()
-
-    delete_translated_images_by_language_code("ko", image_dir)
-    assert not (image_dir / "test.ko.png").exists()
-    assert (image_dir / "test.en.png").exists()
-
-
 def test_map_original_to_translated_exists_and_missing(temp_dir):
     """map_original_to_translated should return the translated path when it exists, else None."""
     root_dir = temp_dir
@@ -240,3 +229,38 @@ def test_map_original_to_translated_exists_and_missing(temp_dir):
     assert mapped_exists is not None
     assert mapped_exists.exists()
     assert mapped_exists == translated_nb.resolve()
+
+
+def create_dummy_image(lang_dir, fname):
+    f = lang_dir / fname
+    f.write_text("dummy image content")
+    return f
+
+
+def test_image_dir_language_subfolders():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        translations_dir = Path(tmpdir) / "translations"
+        image_dir = Path(tmpdir) / "translated_images"
+        langs = ["fr", "de"]
+        reset_translation_directories(translations_dir, image_dir, langs)
+        for lang in langs:
+            assert (translations_dir / lang).is_dir()
+            assert (image_dir / lang).is_dir()
+
+
+def test_delete_translated_images_by_language_code():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        image_dir = Path(tmpdir) / "translated_images"
+        langs = ["fr", "de"]
+        # Setup dirs and dummy files
+        reset_translation_directories(Path(tmpdir) / "translations", image_dir, langs)
+        for lang in langs:
+            lang_dir = image_dir / lang
+            lang_dir.mkdir(parents=True, exist_ok=True)
+            create_dummy_image(lang_dir, f"img1.{lang}.png")
+            create_dummy_image(lang_dir, f"img2.{lang}.jpg")
+            assert any(lang_dir.iterdir())
+        # Delete 'fr' images
+        delete_translated_images_by_language_code("fr", image_dir)
+        assert not (image_dir / "fr").exists()
+        assert (image_dir / "de").is_dir() and any((image_dir / "de").iterdir())
